+++ crates/relayer/src/link/relay_path.rs
@@ -110,8 +110,8 @@ pub struct RelayPath<ChainA: ChainHandle, ChainB: ChainHandle> {
     pending_txs_src: PendingTxs<ChainA>,
     pending_txs_dst: PendingTxs<ChainB>,
 
-    pub max_memo_size: u64,
-    pub max_receiver_size: u64,
+    pub max_memo_size: usize,
+    pub max_receiver_size: usize,
 }
 
 impl<ChainA: ChainHandle, ChainB: ChainHandle> RelayPath<ChainA, ChainB> {
